The swapchain has been separated from the renderer
